Adding velocity command tab completion and wrapping
