+++ Velocity/src/main/java/dev/brighten/antivpn/velocity/VelocityPlugin.java
@@ -10,6 +10,8 @@ import com.velocitypowered.api.plugin.annotation.DataDirectory;
 import com.velocitypowered.api.proxy.ProxyServer;
 import dev.brighten.antivpn.AntiVPN;
 import dev.brighten.antivpn.command.Command;
+import dev.brighten.antivpn.velocity.command.VelocityCommand;
+import dev.brighten.antivpn.velocity.command.VelocityCommandExecutor;
 import lombok.Getter;
 import lombok.val;
 import net.kyori.adventure.text.Component;
@@ -28,7 +30,7 @@ public class VelocityPlugin {
 
     private final ProxyServer server;
     private final Logger logger;
-    private Metrics.Factory metricsFactory;
+    private final Metrics.Factory metricsFactory;
 
     public static VelocityPlugin INSTANCE;
 
@@ -60,40 +62,7 @@ public class VelocityPlugin {
         logger.info("Registering commands...");
         for (Command command : AntiVPN.getInstance().getCommands()) {
             server.getCommandManager().register(server.getCommandManager().metaBuilder(command.name())
-                            .aliases(command.aliases()).build(), (SimpleCommand) invocation -> {
-                        CommandSource sender = invocation.source();
-                        if(!invocation.source().hasPermission("antivpn.command.*")
-                                && !invocation.source().hasPermission(command.permission())) {
-                            invocation.source().sendMessage(Component.text("No permission").toBuilder()
-                                    .color(TextColor.color(255,0,0)).build());
-                            return;
-                        }
-
-                        val children = command.children();
-
-                        String[] args = invocation.arguments();
-                        if(children.length > 0 && args.length > 0) {
-                            for (Command child : children) {
-                                if(child.name().equalsIgnoreCase(args[0]) || Arrays.stream(child.aliases())
-                                        .anyMatch(alias -> alias.equalsIgnoreCase(args[0]))) {
-                                    if(!sender.hasPermission("antivpn.command.*")
-                                            && !sender.hasPermission(child.permission())) {
-                                        invocation.source().sendMessage(Component.text("No permission")
-                                                .toBuilder().color(TextColor.color(255,0,0)).build());
-                                        return;
-                                    }
-                                    sender.sendMessage(LegacyComponentSerializer.builder().character('&').build()
-                                            .deserialize(child.execute(new VelocityCommandExecutor(sender),  IntStream
-                                                    .range(0, args.length - 1)
-                                                    .mapToObj(i -> args[i + 1]).toArray(String[]::new))));
-                                    return;
-                                }
-                            }
-                        }
-
-                        sender.sendMessage(LegacyComponentSerializer.builder().character('&').build()
-                                .deserialize(command.execute(new VelocityCommandExecutor(sender), args)));
-                    });
+                            .aliases(command.aliases()).build(), new VelocityCommand(command));
         }
     }
 }
